(PLEASE READ) Two new optional features:
1. Publish metadata when clients are added or removed.
2. Request timeouts (to defeat replay attacks). This changes the
   register/revoke APIs to require a request-time parameter in the
   JSON body, which will be required to fall within a time interval.

This builds upon the previous commit (which prevents the server's
Ed25519 keypair from being used for remote messages). Since client
registration and revocation both publish to the Chronicle, these
messages are signed by the server's keypair.

As a consequence, if someone builds a higher-level API that reads a
Chronicle to assemble a database of valid clients and their respective
public keys, we do not want to allow someone to abuse a message replay
to introduce a vulnerability in these hypothetical API consumers.